Place NOW below the date; show full weekday name in the header
- Move the jump-to-NOW control to sit centered directly beneath the date pill instead of hanging
  off its left/right edge. This removes the need to track whether the viewed date is in the past
  or future, so the nowOnLeft side decision is gone. The date stays truly centered on the
  viewport center.
- Show the full weekday name in the header top bar (SUNDAY instead of SUN) by switching the
  weekday Intl format from 'short' to 'long'; the month+day stays compact (JUN 28).

Verified in browser: header reads 'SUNDAY ...', NOW renders centered below the date (offset 0).
tsc clean.

+++ components/zero/shell-header.tsx
@@ -25,10 +25,9 @@ function useClock() {
   const time = now
     ? now.toLocaleTimeString([], { hour: "numeric", minute: "2-digit" })
     : ""
-  // Three-letter caps, e.g. "SAT" and "JUN 13" (month first, then day), to keep
-  // the bar compact.
+  // Full weekday in caps (e.g. "SATURDAY") with "JUN 13" (month first, then day).
   const weekday = now
-    ? new Intl.DateTimeFormat("en-US", { weekday: "short" }).format(now).toUpperCase()
+    ? new Intl.DateTimeFormat("en-US", { weekday: "long" }).format(now).toUpperCase()
     : ""
   const monthDay = now
     ? new Intl.DateTimeFormat("en-US", { month: "short", day: "numeric" }).format(now).toUpperCase()
